최종 버전

+++ html/exhibition/db/delete_exhibit.php
@@ -3,14 +3,24 @@
   $table = "exhibit";
   $id = addslashes($_POST['number']);
 
-  $result = query("DELETE FROM $table WHERE id = '$id'");
-  if($result != 0) {
-    echo "
-      <script>
-        alert('삭제되었습니다.');
-        opener.parent.location.reload();
-        window.close();
-      </script>
-    ";
+  $sql = query("SELECT * FROM $table WHERE id = '$id';");
+  $exhibit = $sql->fetch_array();
+  if(unlink("../uploads/".$exhibit['img'])) {
+    $result = query("DELETE FROM $table WHERE id = '$id'");
+    if($result != 0) {
+      echo "
+        <script>
+          alert('삭제되었습니다.');
+          opener.parent.location.reload();
+          window.close();
+        </script>
+      ";
+    }
+    else {
+      echo "DB 삭제 실패";
+    }
+  }
+  else {
+    echo "사진 삭제 실패";
   }
 ?>
